requirements.txt
torch
torchvision
numpy
Pillow
scikit-learn
imagehash
tqdm
git+https://github.com/openai/CLIP.git
